algo_trading_p2/requirements.txt
apache-airflow[amazon,postgres]==2.10.5
SQLAlchemy==1.4.54
psycopg2-binary==2.9.10
numpy==1.26.0
pandas==2.1.1
polars==1.29.0
duckdb==1.2.2
yfinance==0.2.55
